Refatorando
Usando boas práticas herdando classes, é possível sustentar um código mais limpo, organizado e manutenível.
