+++ Pilares da POO/Ecapsulamento-Heranca/src/Exemplo/MSNMessenger.java
@@ -1,20 +1,5 @@
 package Exemplo;
 
-public class MSNMessenger {
-    public void enviarMensagem() {
-        validarConectadoInternet();
-        System.out.println("Eviando mensagem pelo MSN Messenger");
-        salvarHistoricoMensagem();
-    }
-    public void receberMensagem() {
-        System.out.println("Recebendo mensagem pelo MSN Messenger");
-    }
+public class MSNMessenger extends ServicoMensagemInstantanea {
 
-    //métodos privadas, visíveis somente na classe
-    private void validarConectadoInternet() {
-        System.out.println("Validando se está conectado a internet");
-    }
-    private void salvarHistoricoMensagem() {
-        System.out.println("Salvando o histórico da mensagem");
-    }
 }
