+++ Pilares da POO/Ecapsulamento-Heranca/src/Exemplo/ServicoMensagemInstantanea.java
@@ -1,0 +1,23 @@
+package Exemplo;
+
+//a classe MSNMessenger é ou representa
+public class ServicoMensagemInstantanea {
+    public void enviarMensagem() {
+        //primeiro confirmar se esta conectado a internet
+        validarConectadoInternet();
+        System.out.println("Enviando mensagem");
+        //depois de enviada, salva o histórico da mensagem
+        salvarHistoricoMensagem();
+    }
+    public void receberMensagem() {
+        System.out.println("Recebendo mensagem");
+    }
+
+    //métodos privadas, visíveis somente na classe
+    private void validarConectadoInternet() {
+        System.out.println("Validando se está conectado a internet");
+    }
+    private void salvarHistoricoMensagem() {
+        System.out.println("Salvando o histórico da mensagem");
+    }
+}
